Move other placements to preview
[MAILPOET-2811]
