+++ views/form/editor.html
@@ -88,7 +88,6 @@
   'formPlacementPlacementPositionRight': _x('right', 'Placement of a slide in  form, on the left or right side of the page'),
   'formPlacementDelaySeconds': _x('%1s sec', 'times selection should be in the end "30 sec"'),
   'formPlacementSave': _x('Close', 'Text on a button to save and close a form'),
-  'formPlacementOther': _x('Theme widget, block, and others', 'Header in the form placement section'),
   'formPlacementOtherLabel': _x('Other', 'Label in the form placement section (Other form placements)'),
   'addFormWidgetHint': __('You can add this form to a [link]widget area of your theme[/link] (new tab).'),
   'addFormShortcodeHint': __('Or in any page or post as a block, or with this shortcode if you prefer [shortcode].'),
